+++ src/Interval.js
